requirements.txt
discord.py==2.6.4
google-auth==2.45.0
gspread==6.2.1
mongomock==4.2.0.post1
pymongo==4.15.5
pytest==8.3.4
